pyproject.toml
[tool.poetry]
name = "tiny-tiny-diffusion"
version = "0.1.0"
description = ""
authors = ["Kei18 <keisuke.oku18@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ddpm"}]

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.7.2"
numpy = "^1.25.0"
torch = "^2.0.1"
hydra-core = "^1.3.2"
tqdm = "^4.65.0"
pandas = "^2.0.3"
pyyaml = "^6.0"
jupyterlab = ">=3.0.0,<4.0.0a0"
ipympl = "^0.9.3"


[tool.poetry.group.dev.dependencies]
jupyterlab-lsp = "^4.2.0"
isort = "^5.12.0"
black = "^23.3.0"
pre-commit = "^3.3.3"
jupyterlab-code-formatter = "^2.2.1"
python-lsp-server = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
